dev: externalize LockdateForm
